% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquaker-package.R
\docType{package}
\name{earthquaker-package}
\alias{earthquaker}
\alias{earthquaker-package}
\title{earthquaker: R API Wrapper to Retrieve Earthquake Data from the USGS API}
\description{
R API Wrapper to Retrieve Earthquake Data from the USGS API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andypicke/earthquaker}
  \item Report bugs at \url{https://github.com/andypicke/earthquaker/issues}
}

}
\author{
\strong{Maintainer}: Andy Pickering \email{andypicke@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
