% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Function to generate HTML code used in the marker's popup of a leaflet map.
It includes date, location, magnitude and total deaths.
If any of the variables are NA, it is not displayed}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{a data frame from which information is required}
}
\value{
character vector containing HTML code
}
\description{
Function to generate HTML code used in the marker's popup of a leaflet map.
It includes date, location, magnitude and total deaths.
If any of the variables are NA, it is not displayed
}
\examples{
\dontrun{
  eq_clean_data(eq_data()) \%>\% eq_location_clean() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.))

}

}
