% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{eq_data}
\alias{eq_data}
\title{Read the NOAA dataset}
\usage{
eq_data(file.name = file.path(system.file("extdata", package = "earthquakeR"),
  "signif.txt"))
}
\arguments{
\item{file.name}{Path for raw data for NOAA dataset(default - signif.txt)}
}
\value{
A data.frame for NOAA dataset
}
\description{
Read the NOAA dataset
}
\examples{
\dontrun{
  clean_data <- eq_data()
}
}
