% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{CLeans the 'LOCATION_NAME' column as per the requirement (Removing colons and semi-colons)}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{The uncleaned data.frame}
}
\value{
A clean data.frame
}
\description{
CLeans the 'LOCATION_NAME' column as per the requirement (Removing colons and semi-colons)
}
\examples{
\dontrun{
  data <- eq_clean_data(eq_data()) \%>\% eq_location_clean()
}
}
