% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{eq_map}
\alias{eq_map}
\title{Map visualization of earthquakes}
\usage{
eq_map(data, annot_col = "DATE")
}
\arguments{
\item{data}{A data.frame containing the earthquake data}

\item{annot_col}{A column name indicating the column to be used for annotation pop up}
}
\description{
Map visualization of earthquakes
}
\details{
Only requires DATE column in the dataset
}
\examples{
\dontrun{
   eq_clean_data(eq_data()) \%>\% eq_location_clean() \%>\%
  dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\% eq_map()
}

}
