% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake_Plot.R
\name{geom_timeline}
\alias{geom_timeline}
\title{The geom_timeline function}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
This function is used in conjunction with GeomTimeline.
This geom helps to plot a timeline of earthquakes for a given country / countries
with points representing earthquake events, point size representing
earthquake magnitude on richter scale and colour representing number of deaths.
x (the date) is a required aesthetic whereas y (country) is optional.
}
\examples{
\dontrun{
 Plot1 <- eq_rawdata \%>\%
 eq_clean_data() \%>\%
 select(c("Country", "Location_Name", "Date", "Mag", "Latitude", "Longitude", "Deaths")) \%>\%
 filter(Country \%in\% c("India", "China", "Mexico")) \%>\%
 filter(Date>="2000-01-01" & Date<"2010-12-31") \%>\%
 ggplot(aes(x=Date, y=Country, color = Deaths, size = Mag, label=Location_Name)) +
 geom_timeline()
 Plot1
 }
}
