% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake_Plot.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{The geom_timeline_label function}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  n_max = NULL,
  ...
)
}
\arguments{
\item{n_max}{Integer for maximum number of earthquakes to display.}
}
\description{
This function is used in conjunction with GeomTimelinelabel.
Note that user must run geom_timeline first so that we have our points to annotate.
This geom adds a layer to annotate the largest n_max earthquakes with a
vertical line and a label tagging these points.
}
\examples{
\dontrun{
 Plot1 <- eq_rawdata \%>\%
 eq_clean_data() \%>\%
 select(c("Country", "Location_Name", "Date", "Mag", "Latitude", "Longitude", "Deaths")) \%>\%
 filter(Country \%in\% c("India", "China", "Mexico")) \%>\%
 filter(Date>="2000-01-01" & Date<"2010-12-31") \%>\%
 ggplot(aes(x=Date, y=Country, color = Deaths, size = Mag, label=Location_Name)) +
 geom_timeline() \%>\%
 geom_timeline_label(n_max=5) +
 scale_size_continuous(name = "Richter scale value", breaks = c(2,4,6)) +
 scale_color_continuous(name = "# Deaths") +
 theme_bw() +
 theme(legend.position = "right")+
 labs(title = "Earthquakes Summary")
 PLot1
 }
}
