% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Geom for the timeline}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\description{
Geom for the timeline
}
\details{
This is a helper function for geom_timeline and is kept for
        internal use only
}
\keyword{datasets}
