% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Geom  for timeline labels}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}) of length 4.}
\usage{
GeomTimelineLabel
}
\description{
Geom  for timeline labels
}
\details{
This is a helper function for geom_timeline_label and is kept for
        internal use only
}
\keyword{datasets}
