% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake_Plot.R
\name{stat_tll}
\alias{stat_tll}
\title{Stat function for creating data subset to be used for labeling}
\usage{
stat_tll(
  mapping = NULL,
  data = NULL,
  geom = "segment",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  n_max = NULL,
  ...
)
}
\description{
This function is a standard stat function used in conjuction with StatTLL that adds a timeline label layer.
The geom is used as "segment" and n_max is an additional parameter.
}
\examples{
 \dontrun{
 Plot1 <- eq_rawdata \%>\%
 eq_clean_data() \%>\%
 select(c("Country", "Location_Name", "Date", "Mag", "Latitude", "Longitude", "Deaths")) \%>\%
 filter(Country \%in\% c("India", "China", "Mexico")) \%>\%
 filter(Date>="2000-01-01" & Date<"2010-12-31") \%>\%
 ggplot(aes(x=Date, y=Country, color = Deaths, size = Mag, label=Location_Name)) +
 geom_timeline() \%>\% stat_tll(n_max=5)
 Plot1
 }
}
