% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake_Plot.R
\docType{data}
\name{StatTLL}
\alias{StatTLL}
\title{Stat for creating data subset to be used for labeling}
\format{
An object of class \code{StatTLL} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatTLL
}
\description{
A Stat is built using ggproto to be used in conjuction with a geom that adds a timeline label layer.
This stat helps to subset the data to user provided n_max number of earthquakes by magnitude.
}
\examples{
 \dontrun{
 Plot1 <- eq_rawdata \%>\%
 eq_clean_data() \%>\%
 select(c("Country", "Location_Name", "Date", "Mag", "Latitude", "Longitude", "Deaths")) \%>\%
 filter(Country \%in\% c("India", "China", "Mexico")) \%>\%
 filter(Date>="2000-01-01" & Date<"2010-12-31") \%>\%
 ggplot(aes(x=Date, y=Country, color = Deaths, size = Mag, label=Location_Name)) +
 geom_timeline() \%>\% stat_tll(n_max=5)
 PLot1
 }
}
\keyword{datasets}
