% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake_DataClean.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Create Date field using Year, Mo and Dy fields in the raw dataset}
\usage{
eq_clean_data(dataframe)
}
\arguments{
\item{"dataframe"}{is the NOAA dataframe that has the Year, month and day fields to be clubbed and lat/long}
}
\value{
Returns a dataframe with Date field and cleaned location names
}
\description{
NOAA dataset has Year, month and day is separate fields.
\code{eq_clean_data} combines and create a new date field. It also calls the
\code{eq_location_clean} function that creates a cleaned location name. It also converts the
latitude and logitude fields into numeric.
}
\examples{
eq_clean_data(df)

\dontun{
eq_final = eq_rawdata \%>\% #' eq_clean_data()
}
}
