% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquaker_functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean raw data of NOOA dataset}
\usage{
eq_clean_data(raw)
}
\arguments{
\item{raw}{contains the original raw dataset}
}
\value{
This function returns the clean version of the dataset
}
\description{
This function converts time information to Date class,
it converts the longitude and latitude information to numeric
and it applies the eq_location_clean() function to LOCATION_NAME
}
\examples{
\dontrun{
eq_clean_data(raw)
}

}
