% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake_Map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create an earthquake label field for map annotation}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{The dataframe used to create the annotation column.}
}
\value{
Returns a character vector of earthquake details that are used for
map annotations.
}
\description{
This function takes a dataframe (df) and returns a character vector
of HTML column names. Note that the dataframe must have columns
Location_Name, Mag (indicating magnitude) and Deaths.
}
\examples{
\dontrun{
 Map1 <- eq_rawdata \%>\%
 eq_clean_data() \%>\%
 select(c("Country", "Location_Name", "Date", "Mag", "Latitude", "Longitude", "Deaths")) \%>\%
 filter(Country \%in\% c("India", "China", "Mexico")) \%>\%
 filter(Date>="2000-01-01" & Date<"2010-12-31") \%>\%
 mutate(popup_text = eq_create_label(.))
 Map1
}
}
