% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquaker_functions.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create HTML pop up names}
\usage{
eq_create_label(clean)
}
\arguments{
\item{clean}{contains the clean dataset}
}
\value{
This function returns the HTML pop up
}
\description{
The function creates an HTML label for the leaflet map
containing the location cleaned by the eq_location_clean() function,
the magnitude (EQ_PRIMARY), and the total number of deaths (TOTAL_DEATHS)
}
