% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake_DataClean.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean location names of NOAA earthquake dataset}
\usage{
eq_location_clean(dataframe)
}
\arguments{
\item{"dataframe"}{is the NOAA dataframe that has the location name to be cleaned}
}
\value{
Returns a dataframe with cleaned location names
}
\description{
The NOAA dataset location name has a string containing Country and Location names.
\code{eq_location_clean} strips and separates the Country and location names into two fields and
also convert them into title case.
}
\examples{
eq_location_clean(df)

\dontrun{
eq_cleaned <- eq_rawdata \%>\% eq_location_clean()
}
}
