% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquaker_functions.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Parse LOCATION_NAME field of NOOA dataset}
\usage{
eq_location_clean(location_name)
}
\arguments{
\item{location_name}{represents the detailed information about location}
}
\value{
This function returns the clean version of the LOCATION_NAME
}
\description{
This function parses the LOCATION_NAME stripping out the country name
and converting it to title case
}
\examples{
\dontrun{
eq_location_clean(raw$LOCATION_NAME)
}

}
