% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake_Map.R
\name{eq_map}
\alias{eq_map}
\title{Plot a leaflet map with annotations}
\usage{
eq_map(eq_data_clean = NULL, annot_col = "Date")
}
\arguments{
\item{annot_col}{The field from input dataframe to be used for annotation}

\item{df}{The dataframe of data that user wants to plot}
}
\value{
Returns an interactive leaflet map
}
\description{
We plot the earthquake data on map using leaflet and for a subsetted data.
See the leaflet docs (?leaflet) for relevant examples.
The input dataframe requires columns including Longitude, Latitude and Magnitude.
The annotation column ('annot_col') is also a required argument.
}
\examples{
\dontrun{
 Map1 <- eq_rawdata \%>\%
 eq_clean_data() \%>\%
 select(c("Country", "Location_Name", "Date", "Mag", "Latitude", "Longitude", "Deaths")) \%>\%
 filter(Country \%in\% c("India", "China", "Mexico")) \%>\%
 filter(Date>="2000-01-01" & Date<"2010-12-31") \%>\%
 mutate(popup_text = eq_create_label(.)) \%>\%
 eq_map(annot_col = "popup_text")
 Map1
}
}
