% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquaker_functions.R
\name{eq_map}
\alias{eq_map}
\title{Display location of earthquakes}
\usage{
eq_map(clean, annot_col)
}
\arguments{
\item{clean}{contains the clean dataset}

\item{annot_col}{contains the column to be used as pop up window}
}
\value{
This function prints the leaflet map of the earthquakes
}
\description{
The function maps the epicenters (LATITUDE/LONGITUDE) and annotates
each point in a pop up window containing annotation data stored in a
column of the data frame
}
\examples{
\dontrun{
eq_map(clean, annot_col = "popup_text")
}

}
