% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquaker_functions.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Build the Timeline Geom to create earthquakes rapresentation over time}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\value{
This function applies the layer of a GeomTimeline geom
}
\description{
This function applies a ggplot layer of a GeomTimeline geom to a map
}
\examples{
\dontrun{
geom_timeline(to_plot, aes(x = date, y = COUNTRY, i = INTENSITY, d = DEATHS))
}

}
