% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquaker_functions.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Build the TimelineLabel Geom to create labels of earthquakes rapresentation over time}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\value{
This function applies the layer of a GeomTimelineLabel geom
}
\description{
This function applies a ggplot layer of a GeomTimelineLabel geom to a map
}
\examples{
\dontrun{
geom_timeline_label(to_plot, aes(x = date, y = COUNTRY, label = LOCATION_NAME))
}

}
