% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake_Plot.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Defining the GeomTimeline class}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimeline
}
\description{
The ggproto() function is used to construct a new class.
The geom_* function is built as a regular function.
The built geom named geom_timeline() plots the earthquake intensity on a timeline
based on input data range. Optional aesthetics include color, size, and alpha.
The x-aesthetic (required) is a date and an optional y aesthetic is a factor indicating
some stratification in which case multiple time lines will be plotted for
each level of the factor (e.g. country). The points indicate event in time, point size indicate
the earthquake magnitude and colour represent the number of deaths.
}
\examples{
\dontrun{
 Plot1 <- eq_rawdata \%>\%
 eq_clean_data() \%>\%
 select(c("Country", "Location_Name", "Date", "Mag", "Latitude", "Longitude", "Deaths")) \%>\%
 filter(Country \%in\% c("India", "China", "Mexico")) \%>\%
 filter(Date>="2000-01-01" & Date<"2010-12-31") \%>\%
 ggplot(aes(x=Date, y=Country, color = Deaths, size = Mag, label=Location_Name)) +
 geom_timeline()
 PLot1
 }
}
\keyword{datasets}
