% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquaker_functions.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Build the Timeline Geom to create earthquakes rapresentation over time}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\arguments{
\item{required_aes}{contains date, intensity and deaths information}

\item{default_aes}{contains vertical position and info about color}
}
\value{
This function returns a geom to print the earthquakes over time
}
\description{
This function visualizes the times at which earthquakes occur within
certain countries. In addition, it displays the magnitudes (i.e. Richter
scale value) and the number of deaths associated with each earthquake
}
\keyword{datasets}
