% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Earthquake_Plot.R
\docType{data}
\name{GeomTimelinelabel}
\alias{GeomTimelinelabel}
\title{Defines the GeomTimelinelabel class}
\format{
An object of class \code{GeomTimelinelabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimelinelabel
}
\description{
The ggproto() function is used to construct a new class corresponding
to a new geom. The geom_* function is constructed as a regular function which helps
to add annotations to the earthquake data. #' This geom adds a vertical line to
each data point with a text annotation (e.g. the location of the earthquake) attached
to each line. There is an option for user to subset to n_max number of earthquakes
where we take the n_max largest (by magnitude) earthquakes. Aesthetics are x, which
is the date of the earthquake and label which takes the column name from which
annotations will be obtained. Please note that user must run geom_timeline first so
that we have our points to annotate.
}
\examples{
 \dontrun{
 Plot1 <- eq_rawdata \%>\%
 eq_clean_data() \%>\%
 select(c("Country", "Location_Name", "Date", "Mag", "Latitude", "Longitude", "Deaths")) \%>\%
 filter(Country \%in\% c("India", "China", "Mexico")) \%>\%
 filter(Date>="2000-01-01" & Date<"2010-12-31") \%>\%
 ggplot(aes(x=Date, y=Country, color = Deaths, size = Mag, label=Location_Name)) +
 geom_timeline() \%>\%
 geom_timeline_label(n_max=5) +
 scale_size_continuous(name = "Richter scale value", breaks = c(2,4,6)) +
 scale_color_continuous(name = "# Deaths") +
 theme_bw() +
 theme(legend.position = "right")+
 labs(title = "Earthquakes Summary")
 Plot1
 }
}
\keyword{datasets}
