% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquaker_functions.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Build the TimelineLabel Geom to create labels over earthquakes rapresentation over time}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimelineLabel
}
\arguments{
\item{required_aes}{contains x, y positions and name of the earthquake}

\item{default_aes}{}
}
\value{
This function returns a geom to print the labels of earthquakes over time
}
\description{
This function adds labels of the earthquakes on a Timeline representation
}
\keyword{datasets}
