% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_earthquakes.R
\name{get_earthquakes}
\alias{get_earthquakes}
\title{Retrieve earthquake data from USGS API}
\usage{
get_earthquakes(
  format = "csv",
  starttime = Sys.Date() - 15,
  endtime = Sys.Date(),
  min_magnitude = 0,
  alert_level = c(NA, "green", "yellow", "orange", "red")
)
}
\arguments{
\item{format}{Format to return. Default is csv}

\item{starttime}{Starting date/time in yyyy-mm-dd format (default = 15 days ago)}

\item{endtime}{Ending date/time in yyyy-mm-dd format (default = today)}

\item{min_magnitude}{Minimum magnitude (default = 0)}

\item{alert_level}{(optional) PAGER alert level. c(NA, "green", "yellow", "orange", "red")}
}
\value{
df Dataframe of earthquakes data with requested parameters.
}
\description{
Retrieve earthquake data from USGS API
}
